requirement.txt
absl-py==1.2.0
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
albumentations==1.1.0
alembic==1.4.1
altair==4.1.0
altgraph==0.17
aniso8601==9.0.1
anyio==3.3.0
appdirs==1.4.4
argcomplete==1.10.3
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
astroid==2.9.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bayesian-optimization==1.2.0
beautifulsoup4==4.10.0
bleach==4.0.0
blinker==1.4
blis==0.7.4
Boruta==0.3
Bottleneck @ file:///C:/Users/1768252/Bottleneck-1.3.2-cp38-cp38-win_amd64.whl
bs4==0.0.1
cachetools==4.2.2
catalogue==1.0.0
catboost==1.0.6
cchardet==2.1.7
certifi==2021.5.30
certipy==0.1.3
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==1.6.0
cmake==3.22.4
colorama==0.4.4
colorlover==0.3.0
commonmark==0.9.1
compressed-rtf==1.0.6
comtypes==1.1.10
cryptography==35.0.0
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
databricks-cli==0.15.0
datatable==1.0.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
docker==5.0.0
docopt==0.6.2
docx2txt==0.8
dython==0.7.1.post3
ebcdic==1.1.1
EbookLib==0.17.1
elastic-transport==8.1.1
elasticsearch==8.1.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
entrypoints==0.3
et-xmlfile==1.1.0
extended-maths==0.1.1
extract-msg==0.28.7
face-recognition-models==0.3.0
fake-useragent==0.1.11
fastapi==0.68.1
Fiona @ file:///C:/Users/1768252/Downloads/Fiona-1.8.21-cp38-cp38-win_amd64.whl
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
frozenlist==1.3.0
funcy==1.16
future==0.18.2
gast==0.4.0
gensim==3.8.3
geographiclib==1.52
geopandas==0.10.2
geopy==2.2.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-trans-new==1.1.9
googletransx==2.4.2
graphviz==0.20
greenlet==1.1.1
grpcio==1.41.1
gtrending==0.3.5
h11==0.12.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
hstspreload==2021.12.1
htmlmin==0.1.12
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
hyperframe==6.0.1
idna==2.8
igramscraper==0.3.2
ImageHash==4.2.1
imageio==2.13.5
IMAPClient==2.1.0
imbalanced-learn==0.7.0
importlib-metadata==4.11.3
instagram-scraper==1.10.3
ipykernel==6.1.0
ipynb-py-convert==0.4.6
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
Js2Py==0.71
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyter-telemetry==0.1.0
jupyterhub==2.2.0
jupyterlab==3.1.6
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.0
jupyterlab-widgets==1.0.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kmodes==0.11.0
lazy-object-proxy==1.7.1
libclang==14.0.1
libretranslatepy==2.1.1
lifeeasy==2.3.3
lightgbm==3.2.1
linecache2==1.0.0
llvmlite==0.36.0
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
missingno==0.5.0
mistune==0.8.4
mlflow==1.19.0
mlxtend==0.18.0
MouseInfo==0.1.3
mpl-finance==0.10.1
mplfinance==0.12.8b9
multidict==6.0.2
multimethod==1.4
munch==2.5.0
murmurhash==1.0.5
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==5.6.1
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.3
notebook-as-pdf==0.5.0
numba==0.53.1
numexpr==2.7.3
numpy==1.21.6
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.60
opencv-python-headless==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.0
pandas==1.4.3
pandas-datareader==0.10.0
pandas-profiling==3.0.0
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.0
patsy==0.5.1
pdfminer.six==20191110
pefile==2021.5.24
phik==0.12.0
pickleshare==0.7.5
Pillow==9.0.0
pipwin==0.5.2
plac==1.1.3
platformdirs==2.4.0
plotly==5.5.0
preshed==3.0.5
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.9.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycares==4.1.2
pycaret==2.3.3
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.12.0
pydantic==1.8.2
pydeck==0.6.2
pyee==8.2.2
PyGetWindow==0.0.9
Pygments==2.9.0
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyjsparser==2.7.1
pyLDAvis==3.2.2
pylint==2.12.2
pymongo==3.10.1
Pympler==1.0.1
PyMsgBox==1.0.9
pynndescent==0.5.4
pyod==0.9.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypi==2.1
pyppeteer==1.0.2
PyPrind==2.11.3
pyproj==3.3.0
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
python-json-logger==2.0.2
python-multipart==0.0.5
python-pptx==0.6.21
python-slugify==3.0.2
python-socks==2.0.3
pytube==11.0.1
pytweening==1.0.4
pytz==2021.1
PyWavelets==1.1.1
pywin32==302
pywin32-ctypes==0.2.0
pywinauto==0.6.8
pywinpty==1.1.3
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
qudida==0.0.4
querystring-parser==1.2.4
regex==2021.8.3
requests==2.27.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==12.5.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
schedule==1.1.0
scikit-image==0.19.2
scikit-learn==1.1.1
scikit-plot==0.3.7
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
Shapely==1.8.1.post1
six==1.12.0
sklearn==0.0
smart-open==5.2.0
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy-legacy==3.0.8
SQLAlchemy==1.4.23
sqlparse==0.4.1
srsly==1.0.5
starlette==0.14.2
statsmodels==0.12.2
streamlit==1.10.0
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.11.0
testpath==0.5.0
text-unidecode==1.2
textblob==0.15.3
textract==1.6.4
thinc==7.4.5
threadpoolctl==2.2.0
tifffile==2022.5.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.1
traceback2==1.4.0
traitlets==5.0.5
translate==3.6.1
trio==0.19.0
trio-websocket==0.9.2
twint==2.1.20
typer==0.3.2
typing_extensions==4.0.1
tzdata==2021.1
tzlocal==3.0
umap-learn==0.5.1
unicorn==1.0.3
urllib3==1.26.8
uvicorn==0.18.2
validators==0.18.2
visions==0.7.1
waitress==2.0.0
wasabi==0.8.2
watchdog==2.1.4
watchfiles==0.15.0
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==10.2
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
wsproto==1.0.0
XlsxWriter==3.0.2
yarl==1.7.2
yellowbrick==1.3.post1
zipp==3.8.0
